std140 and padding as fucking vec4s to the rescue.

+++ Examples_3/Unit_Tests/src/15_Transparency/15_Transparency.cpp
@@ -975,10 +975,8 @@ class Transparency: public IApp
 
 		const size_t size = frustumGrid.lightCounts.size();
 		for (size_t i = 0; i < size; i++) {
-			vec4 lightCount = { frustumGrid.lightCounts[i], 0, 0, 1 };
+			vec4 lightCount = { (float)frustumGrid.lightCounts[i], 0, 0, 1 };
 			gMaterialUniformData.lightCounts[i] = lightCount;
-			//volatile int dog = 5;
-			//gMaterialUniformData.lightCounts[i] = gMaterialUniformData.lightCounts[i].setX(frustumGrid.lightCounts[i]);
 		}
 
 		//memcpy(/*gHeatmapUniformData*/gMaterialUniformData.lightCounts, frustumGrid.lightCounts.data(), sizeof(int) * frustumGrid.lightCounts.size());
